requirements.txt
dspy>=3.1.0
gymnasium[classic-control]
pygame
Pillow
